main's args & init file parsing..(read_more)
i am parsing init file reading the words of each row and adding them into 
an ngram (implemented as a list of words).
that ngram is to be inserted into the Trie.

I also made a sloppy makefile.

parsing of init.txt was kind of easy and i did it in main.c
so that i could insert the ngram in trie before the reading 
of the next line.